refactor(api-service): clean up Phase 5/6 architecture
- Extract AgentInboundHandler from ChatSdkService — ChatSdkService now purely
  manages SDK lifecycle + delivery, inbound orchestration is a separate service
- Promote resolve to first-class lifecycle action on reply payload, no longer
  a signal — allows resolving without replying (timeouts, button clicks)
- Remove redundant agentId from BridgeExecutorParams and HandleAgentReplyCommand
  — JWT claims are the sole authority for routing
- Merge duplicate ReplyContentDto/UpdateContentDto into TextContentDto
- Extract getPrimaryChannel helper to validate channels[0] once
- Remove unused imports and dead injections

Made-with: Cursor
