+++ apps/api/src/app/agents/agents.module.ts
@@ -12,6 +12,7 @@ import { AgentsController } from './agents.controller';
 import { AgentsWebhookController } from './agents-webhook.controller';
 import { AgentConversationService } from './services/agent-conversation.service';
 import { AgentCredentialService } from './services/agent-credential.service';
+import { AgentInboundHandler } from './services/agent-inbound-handler.service';
 import { AgentSubscriberResolver } from './services/agent-subscriber-resolver.service';
 import { BridgeExecutorService } from './services/bridge-executor.service';
 import { ChatSdkService } from './services/chat-sdk.service';
@@ -29,6 +30,7 @@ import { USE_CASES } from './usecases';
     AgentCredentialService,
     AgentSubscriberResolver,
     AgentConversationService,
+    AgentInboundHandler,
     BridgeExecutorService,
     ChatSdkService,
   ],
