+++ apps/api/src/app/agents/services/agent-inbound-handler.service.ts
@@ -1,0 +1,101 @@
+import { Injectable } from '@nestjs/common';
+import { PinoLogger } from '@novu/application-generic';
+import { ConversationParticipantTypeEnum, SubscriberRepository } from '@novu/dal';
+import type { Message, Thread } from 'chat';
+import { AgentEventEnum } from '../dtos/agent-event.enum';
+import { ResolvedPlatformConfig } from './agent-credential.service';
+import { AgentConversationService } from './agent-conversation.service';
+import { AgentSubscriberResolver } from './agent-subscriber-resolver.service';
+import { BridgeExecutorService } from './bridge-executor.service';
+
+@Injectable()
+export class AgentInboundHandler {
+  constructor(
+    private readonly logger: PinoLogger,
+    private readonly subscriberResolver: AgentSubscriberResolver,
+    private readonly conversationService: AgentConversationService,
+    private readonly bridgeExecutor: BridgeExecutorService,
+    private readonly subscriberRepository: SubscriberRepository
+  ) {}
+
+  async handle(
+    agentId: string,
+    config: ResolvedPlatformConfig,
+    thread: Thread,
+    message: Message,
+    event: AgentEventEnum
+  ): Promise<void> {
+    const subscriberId = await this.subscriberResolver
+      .resolve({
+        environmentId: config.environmentId,
+        organizationId: config.organizationId,
+        platform: config.platform,
+        platformUserId: message.author.userId,
+        integrationIdentifier: config.integrationIdentifier,
+      })
+      .catch((err) => {
+        this.logger.warn(err, `[agent:${agentId}] Subscriber resolution failed, continuing without subscriber`);
+
+        return null;
+      });
+
+    const participantId = subscriberId ?? `${config.platform}:${message.author.userId}`;
+    const participantType = subscriberId
+      ? ConversationParticipantTypeEnum.SUBSCRIBER
+      : ConversationParticipantTypeEnum.PLATFORM_USER;
+
+    const conversation = await this.conversationService.createOrGetConversation({
+      environmentId: config.environmentId,
+      organizationId: config.organizationId,
+      agentId,
+      platform: config.platform,
+      integrationId: config.integrationId,
+      platformThreadId: thread.id,
+      participantId,
+      participantType,
+      platformUserId: message.author.userId,
+      firstMessageText: message.text,
+    });
+
+    await this.conversationService.persistInboundMessage({
+      conversationId: conversation._id,
+      platform: config.platform,
+      integrationId: config.integrationId,
+      platformThreadId: thread.id,
+      senderId: participantId,
+      senderName: message.author.fullName,
+      content: message.text,
+      platformMessageId: message.id,
+      environmentId: config.environmentId,
+      organizationId: config.organizationId,
+    });
+
+    await thread.startTyping();
+
+    const serializedThread = thread.toJSON() as unknown as Record<string, unknown>;
+    await this.conversationService.updateChannelThread(
+      config.environmentId,
+      config.organizationId,
+      conversation._id,
+      thread.id,
+      serializedThread
+    );
+
+    const [subscriber, history] = await Promise.all([
+      subscriberId
+        ? this.subscriberRepository.findBySubscriberId(config.environmentId, subscriberId)
+        : Promise.resolve(null),
+      this.conversationService.getHistory(config.environmentId, conversation._id),
+    ]);
+
+    await this.bridgeExecutor.execute({
+      event,
+      config,
+      conversation,
+      subscriber,
+      history,
+      message,
+      thread,
+    });
+  }
+}
